requirements.txt
# TAISE-Agent v0.5 Dependencies
fastapi>=0.109.0
uvicorn>=0.27.0
httpx>=0.26.0
pyyaml>=6.0.1
jinja2>=3.1.3
anthropic>=0.45.0
openai>=1.12.0
python-dotenv>=1.0.0
pydantic>=2.6.0

# v0.2 additions
python-telegram-bot>=21.0       # Telegram adapter for OpenClaw
